requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
filelock==3.15.4
freetype-py==2.3.0
frozenlist==1.4.1
fsspec==2024.6.1
geoip2==4.8.0
greenlet==3.0.3
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.8
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
maxminddb==2.6.2
mkl-service==2.4.1
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
pandas==2.2.2
pillow==10.4.0
pip==24.2
pycairo==1.26.1
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.1
reportlab==4.2.2
requests==2.32.3
rlPyCairo==0.2.0
scikit-learn==1.5.1
scipy==1.14.1
setuptools==72.2.0
six==1.16.0
sympy==1.13.2
threadpoolctl==3.5.0
torch==2.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wheel==0.44.0
win-inet-pton==1.1.0
yarl==1.9.6
zstandard==0.23.0
